* EOAccess/EODatabaseContext.m -initializeObject:withGlobalID:editingContext: remove exessive checking -initializeObject:row:entity:editingContext: reformat, less IMPs, less logs * EOControl/EOFault.h + retainCount unsigned -> NSUInteger * EOControl/EOFault.m + retainCount unsigned -> NSUInteger + clearFault: raise if argument is no fault rewrote retain behaviour it makes no sense to free objects here and use the pointers this would mess up the whole memory and crash.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gdl2/trunk@30886 72102866-910b-0410-8b05-ffd578937521

+++ EOControl/EOFault.h
@@ -65,7 +65,7 @@
 + (id)retain;
 + (void)release;
 + (id)autorelease;
-+ (unsigned)retainCount;
++ (NSUInteger)retainCount;
 
 + (BOOL)isKindOfClass: (Class)aClass;
 + (void)doesNotRecognizeSelector: (SEL)selector;
